Remove container map and move routing components folder.
